+++ include/PerlinNoise.h
@@ -18,11 +18,11 @@ public:
 	// Generate a new permutation vector based on the value of seed
 	PerlinNoise(unsigned int seed);
 	// Get a noise value, for 2D images z can have any value
-	double noise(double x, double y, double z);
+	float noise(float x, float y, float z);
 private:
-	double fade(double t);
-	double lerp(double t, double a, double b);
-	double grad(int hash, double x, double y, double z);
+	float fade(float t);
+	float lerp(float t, float a, float b);
+	float grad(int hash, float x, float y, float z);
 };
 
 #endif
